feat(web): extract brief email helpers into brief-email-lib
Move weekly brief HTML/text rendering into a dedicated -lib module with
expanded Vitest coverage while preserving the public @/lib/brief-email
import surface.

Closes #4291
